move HitShape to Bits
