+++ src/Blitter.Blocks/Playfield/PlayField2D.cs
@@ -1,4 +1,5 @@
 using System.Collections.Immutable;
+using Blitter.Bits;
 
 namespace Blitter.Blocks;
